+++ docs/.vuepress/clientAppEnhance.js
@@ -1,0 +1,6 @@
+import { h } from 'vue';
+import { defineClientAppEnhance } from '@vuepress/client';
+import Aurora from './components/Aurora'
+export default defineClientAppEnhance(({ app, router }) => {
+    app.component("Aurora",Aurora)
+});