+++ docs/.vuepress/theme/lib/client/components/About.vue
@@ -124,6 +124,7 @@ export default defineComponent({
   },
 
   created() {
+
     //如果手机端侧边栏打开的，那么就关闭
     if (this.$store.state.openMobileSidebar) {
       this.$store.commit("setOpenMobileSidebar",{